Move RunModes to analyser api
